Day 09
Upgrade Intcode computer to handle large numbers and relative positions

-- BOOST doesn't yet work

+++ src/main/kotlin/day05/ThermalEnvironmentSupervisionTerminal.kt
@@ -4,7 +4,7 @@ import intcode.IntcodeComputer
 
 class ThermalEnvironmentSupervisionTerminal {
 
-    fun runTEST(input: Int) {
+    fun runTEST(input: Long) {
         val program = IntcodeComputer.readInput("src/main/resources/day05/input.txt")
         val intcodeComputer = IntcodeComputer(program.toMutableList(), mutableListOf(input))
         val output = intcodeComputer.calculateIntcode()
